+++ app/assets/javascripts/mixins/contextual_serach_operands.js
@@ -1,0 +1,8 @@
+const Operands = {
+  requester: 'requested_by_name:',
+  owner: 'owned_by_name:',
+  initials: 'owned_by_initials:',
+  type: 'story_type:',
+};
+
+module.exports = Operands;